Use config file to find data paths.
Rather than hard-coding paths to the raw data files, I now check for
a base data directory as specified in CONFIG.cfg. Since the path to this
is set relative to the build file that looks for it, it *should* avoid
absolute import issues, but this might need to be revisited.

As a result, I also refactored the database code to accept the config
file contents, and interpret the path, rather than having the complete
path be passed as a parameter in the Database __init__ method.

Finally, I (temporarily) abandoned trying to use multiprocessing for
parsing hetionet edges. While testing the build scripts, I instead
just totally skip hetionet edges, which shouldn't cause major issues
while working on the other database imports.