Finish implementing adjacency matrix code.
Currently, only sparse adjacency marices are supported, in the Scipy
List-of-Lists format. Since the graph is directed, the adjacency matrix
is not symmetric, but we should probably provide a version of the
adjacency matrix code that treats the graph as undirected, for spectral
computations.

I wrote several new tests for the Graph, but it seems the fixture isn't
working properly - currently, it recomputes the networkx object for each
test that uses the metrics module, which it should only do once.


Former-commit-id: dcd26300abe7eb2314d6a3b97ef28c7f436f6b86