Add AOP subgraph builder algorithm.
This algorithm accepts an AOP title, which it uses to retrieve the
database node from the graph database. This node is used to walk a subgraph
using a depth-first approach, constructing a subnetwork bounded by AOP-related
edge types (e.g., Neo4j edge 'labels'). Documentation will be provided in
the near future describing the algorithm formally.