Formalize subgraph.
Begin to build out the implementation of subgraphs. An important
characteristic is that they are *not* a descendant of Graph. However,
they are always derived from a currently instantiated graph. Also, they
are implemented as a view into the graph from which they are derived,
meaning data is NOT copied (a substantial advantage)


Former-commit-id: 1a4fad50057032bc4d6f64b9601a59cd86ef469c