Overhaul graph database (major update).
The previous method for building the graph database was unsustainable
for long-term use, so I created a more modular build process, as is now
shown in comptox_ai/build/db.py. This should probably be split up into
several files, and there are important tweaks that need to be
incorporated, but for the most part it demonstrates how the graph
database *should* be built for the most part.

I've added a number of new properties and restructured much of the
ontology, so those changes are also demonstrated in comptox.rdf.