Enable connection to remote database.
Note that we have only tested this on the public version of the
ComptoxAI graph database, which is currently hosted at
165.123.13.192. Unfortunately, it DOESN'T work if you use the
comptox.ai domain as the hostname of the neo4j database (e.g., via
`neo4j://neo4j.comptox.ai` or `bolt://neo4j.comptox.ai`). This is
relatively high-priority to figure out, and should be fixed ASAP, as
it will create a confusing scenario for new users.