[#51] Refer to GitHub issue...
This update adds two GitHub Action workflows: doc-build.yml and react-build.yml.

doc-build.yml is used to create the documentation page using Sphinx.

react-build.yml is used to create the Comptox AI app.

conf.py file located at /docs/source was copied to /docs in order to use the file when building with Sphinx.

setup.py was modified to add dependencies to extras_regular that are needed to build with Sphinx.