cd: Add GitHub action to deploy website to AWS
[#63] Refer to GitHub issue…

This update creates a GitHub action that deploys ComptoxAI website to
AWS EC2 instance on GitHub release.
The action uses a pre-made GitHub action from
easingthemes/ssh-deploy@v2.
The ssh keys needed are called from GitHub secret.

In the future, there should be a review on whether changing the
authentication method to AWS role is a safer method.
In this case, the action must be built from scratch rather than using an
external one as it uses different parameters.