fix: Removed Py2neo-related code
[#93] Refer to GitHub issue...

This commit removed all code references to the Py2neo package, which
were previously removed from the project's dependencies. The changes
included:

- Deleted Python code that imported and referenced Py2neo.
- Commented out references to the `Graph` class in `graph/graph.py` as
it was based on Py2neo. The decision to comment out instead of deleting
was made for future reference purposes.
- Autoformatted several Python files that hadn't been previously styled
according to the project's style guide.
- Removed the Py2neo package from `environment.yml`.

The current commit successfully passed the Pytest after modification.