docs: Simplify GitHub Action badge text in README
[#89] Refer to GitHub issue...

This update aims to address the issue with GitHub's README page not
displaying alt text correctly for GitHub Action badge status. The README
Markdown for badge status has been simplified to make it more
straightforward. While this change may not guarantee a fix on the GitHub
page, it serves as an attempt to improve badge rendering. The update
doesn't alter any functionality from the previous version and will be
retained regardless of its impact on the GitHub README page.